+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Timestepping: add `LIA` option to `init`, enabling simulations of the local
   induction approximation.
 
+- Timestepping: add `MaximumTimestep` adaptivity criterion. This criterion is
+  implicitly applied when other criteria are passed which allow "infinite"
+  timesteps (the case of `AdaptBasedOnVelocity`).
+
 - Write filament parametrisation (knots) to VTKHDF files in `write_vtkhdf`.
   This can be disabled by passing `parametrisation = false`.
   Also, if the parametrisation is present in a VTKHDF file, then `read_vtkhdf`